add MainFrame and fix BanAn
