--- BanAn/nbproject/genfiles.properties
@@ -1,8 +1,0 @@
-build.xml.data.CRC32=c844d166
-build.xml.script.CRC32=1c6ff257
-build.xml.stylesheet.CRC32=f85dc8f2@1.99.0.48
-# This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
-# Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=c844d166
-nbproject/build-impl.xml.script.CRC32=233726c9
-nbproject/build-impl.xml.stylesheet.CRC32=d549e5cc@1.99.0.48
